Implement mission joining action
